requirements.txt
crewai
crewai[tools]
google-genai
google-play-scraper
pandas
python-dotenv